requirements.txt
fastapi
uvicorn
pydantic
torch
tqdm
pyyaml
aioredis  # optional, for Redis-based message bus
requests
deepgram-sdk==2.12.0
pytest
pytest-asyncio
httpx 